refactor: move inline tests to sidecars
