+++ src/notifications/dispatcher.rs
@@ -449,273 +449,5 @@ pub(crate) fn spawn_dispatcher(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::StorageConfig;
-    use crate::db::init_pool;
-    use crate::db::notifications::outbox_insert;
-
-    fn open_test_db() -> rusqlite::Connection {
-        let conn = rusqlite::Connection::open_in_memory().unwrap();
-        conn.execute_batch(
-            "CREATE TABLE notifications_outbox (
-                 id INTEGER PRIMARY KEY AUTOINCREMENT,
-                 dedup_key TEXT NOT NULL,
-                 rule_id TEXT NOT NULL,
-                 severity TEXT NOT NULL,
-                 hostname TEXT NOT NULL,
-                 title TEXT NOT NULL,
-                 body TEXT NOT NULL,
-                 apprise_urls_json TEXT NOT NULL,
-                 apprise_tags TEXT,
-                 enqueued_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%fZ','now')),
-                 next_attempt_at TEXT NOT NULL,
-                 attempt_count INTEGER NOT NULL DEFAULT 0,
-                 last_status_code INTEGER,
-                 last_error TEXT,
-                 status TEXT NOT NULL DEFAULT 'pending'
-                     CHECK (status IN ('pending','sent','dead','dropped'))
-             );
-             CREATE UNIQUE INDEX IF NOT EXISTS idx_outbox_dedup_pending
-                 ON notifications_outbox(dedup_key) WHERE status = 'pending';
-             CREATE TABLE notification_firings (
-                 id INTEGER PRIMARY KEY AUTOINCREMENT,
-                 outbox_id INTEGER NOT NULL,
-                 rule_id TEXT NOT NULL,
-                 severity TEXT NOT NULL,
-                 hostname TEXT NOT NULL,
-                 fired_at TEXT NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%fZ','now')),
-                 status_code INTEGER,
-                 notes TEXT,
-                 dedup_key TEXT NOT NULL DEFAULT ''
-             );
-             CREATE TABLE error_signatures (
-                 signature_hash TEXT NOT NULL,
-                 normalizer_version INTEGER NOT NULL,
-                 template TEXT NOT NULL,
-                 sample_message TEXT NOT NULL,
-                 sample_hostname TEXT NOT NULL,
-                 sample_app_name TEXT,
-                 severity TEXT NOT NULL,
-                 first_seen_at TEXT NOT NULL,
-                 last_seen_at TEXT NOT NULL,
-                 total_count INTEGER NOT NULL DEFAULT 0,
-                 acknowledged_at TEXT,
-                 acknowledged_by TEXT,
-                 PRIMARY KEY (signature_hash, normalizer_version)
-             );",
-        )
-        .unwrap();
-        conn
-    }
-
-    #[test]
-    fn backoff_schedule_is_increasing() {
-        let delays: Vec<_> = (0u8..5)
-            .map(|i| {
-                let s = backoff_next_attempt_at(i);
-                chrono::DateTime::parse_from_rfc3339(&s).unwrap()
-            })
-            .collect();
-        for window in delays.windows(2) {
-            assert!(
-                window[1] >= window[0],
-                "backoff should be non-decreasing: {:?} >= {:?}",
-                window[1],
-                window[0]
-            );
-        }
-    }
-
-    #[test]
-    fn severity_to_notify_type_mapping() {
-        assert_eq!(severity_to_notify_type("crit"), NotifyType::Failure);
-        assert_eq!(severity_to_notify_type("critical"), NotifyType::Failure);
-        assert_eq!(severity_to_notify_type("warning"), NotifyType::Warning);
-        assert_eq!(severity_to_notify_type("err"), NotifyType::Warning);
-        assert_eq!(severity_to_notify_type("notice"), NotifyType::Info);
-        assert_eq!(severity_to_notify_type("info"), NotifyType::Info);
-    }
-
-    #[test]
-    fn signature_ack_check_distinguishes_missing_pending_and_acknowledged() {
-        let conn = open_test_db();
-        let version = crate::app::error_detection::NORMALIZER_VERSION;
-
-        assert!(!is_signature_acked(&conn, "missing", version).unwrap());
-
-        conn.execute(
-            "INSERT INTO error_signatures (
-                 signature_hash, normalizer_version, template, sample_message,
-                 sample_hostname, severity, first_seen_at, last_seen_at, acknowledged_at
-             ) VALUES (?1, ?2, 'template', 'sample', 'host1', 'err',
-                       '2026-06-13T00:00:00.000Z', '2026-06-13T00:00:00.000Z', NULL)",
-            rusqlite::params!["pending", version],
-        )
-        .unwrap();
-        assert!(!is_signature_acked(&conn, "pending", version).unwrap());
-
-        conn.execute(
-            "INSERT INTO error_signatures (
-                 signature_hash, normalizer_version, template, sample_message,
-                 sample_hostname, severity, first_seen_at, last_seen_at, acknowledged_at
-             ) VALUES (?1, ?2, 'template', 'sample', 'host1', 'err',
-                       '2026-06-13T00:00:00.000Z', '2026-06-13T00:00:00.000Z',
-                       '2026-06-13T00:01:00.000Z')",
-            rusqlite::params!["acked", version],
-        )
-        .unwrap();
-        assert!(is_signature_acked(&conn, "acked", version).unwrap());
-        assert!(!is_signature_acked(&conn, "acked", version + 1).unwrap());
-    }
-
-    #[test]
-    fn outbox_row_dedup_suppressed() {
-        let conn = open_test_db();
-        // Insert a firing first (simulating a previous delivery)
-        let params = crate::db::notifications::OutboxInsertParams {
-            dedup_key: "oom_kill:host1:ts".to_string(),
-            rule_id: "oom_kill".to_string(),
-            severity: "critical".to_string(),
-            hostname: "host1".to_string(),
-            title: "OOM".to_string(),
-            body: "body".to_string(),
-            apprise_urls_json: "[]".to_string(),
-            next_attempt_at: "2000-01-01T00:00:00.000Z".to_string(),
-        };
-        outbox_insert(&conn, &params).unwrap();
-        let id: i64 = conn
-            .query_row("SELECT id FROM notifications_outbox LIMIT 1", [], |r| {
-                r.get(0)
-            })
-            .unwrap();
-        // Insert a firing within the dedup window
-        firings_insert(
-            &conn,
-            FiringInsertParams {
-                outbox_id: id,
-                rule_id: "oom_kill",
-                severity: "critical",
-                hostname: "host1",
-                status_code: Some(200),
-                notes: None,
-                dedup_key: "oom_kill:host1:ts",
-            },
-        )
-        .unwrap();
-
-        let is_dedup =
-            firings_recent_dedup_check(&conn, "oom_kill", "host1", "oom_kill:host1:ts", 3600)
-                .unwrap();
-        assert!(
-            is_dedup,
-            "should detect existing firing within dedup window"
-        );
-    }
-
-    #[test]
-    fn dead_letter_after_max_retries() {
-        let conn = open_test_db();
-        let params = crate::db::notifications::OutboxInsertParams {
-            dedup_key: "test:dl".to_string(),
-            rule_id: "oom_kill".to_string(),
-            severity: "critical".to_string(),
-            hostname: "host1".to_string(),
-            title: "Title".to_string(),
-            body: "Body".to_string(),
-            apprise_urls_json: "[]".to_string(),
-            next_attempt_at: "2000-01-01T00:00:00.000Z".to_string(),
-        };
-        outbox_insert(&conn, &params).unwrap();
-        let id: i64 = conn
-            .query_row("SELECT id FROM notifications_outbox LIMIT 1", [], |r| {
-                r.get(0)
-            })
-            .unwrap();
-
-        // Simulate 8 failed retries
-        let max_retries: u8 = 8;
-        for attempt in 0u8..max_retries {
-            let next_at = backoff_next_attempt_at(attempt);
-            outbox_schedule_retry(&conn, id, &next_at, "timeout", None).unwrap();
-        }
-
-        let attempt_count: i64 = conn
-            .query_row(
-                "SELECT attempt_count FROM notifications_outbox WHERE id = ?1",
-                rusqlite::params![id],
-                |r| r.get(0),
-            )
-            .unwrap();
-
-        assert!(
-            attempt_count >= max_retries as i64,
-            "attempt_count={attempt_count} should be >= max_retries={max_retries}"
-        );
-
-        // Mark dead
-        outbox_mark_dead(&conn, id, None, "exhausted").unwrap();
-        let status: String = conn
-            .query_row(
-                "SELECT status FROM notifications_outbox WHERE id = ?1",
-                rusqlite::params![id],
-                |r| r.get(0),
-            )
-            .unwrap();
-        assert_eq!(status, "dead");
-    }
-
-    #[tokio::test]
-    async fn dispatch_cycle_drops_pending_row_when_no_apprise_urls_exist() {
-        let dir = tempfile::tempdir().unwrap();
-        let pool = Arc::new(
-            init_pool(&StorageConfig::for_test(
-                dir.path().join("notifications.db"),
-            ))
-            .unwrap(),
-        );
-        let conn = pool.get().unwrap();
-        outbox_insert(
-            &conn,
-            &crate::db::notifications::OutboxInsertParams {
-                dedup_key: "no-url:test".to_string(),
-                rule_id: "oom_kill".to_string(),
-                severity: "critical".to_string(),
-                hostname: "host1".to_string(),
-                title: "Title".to_string(),
-                body: "Body".to_string(),
-                apprise_urls_json: "[]".to_string(),
-                next_attempt_at: "2000-01-01T00:00:00.000Z".to_string(),
-            },
-        )
-        .unwrap();
-        drop(conn);
-
-        let dispatched = run_dispatch_cycle(
-            Arc::clone(&pool),
-            Arc::new(Semaphore::new(1)),
-            &AppriseClient::new("http://127.0.0.1:1"),
-            &NotificationsConfig::default(),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(dispatched, 0);
-        let conn = pool.get().unwrap();
-        let (status, error): (String, Option<String>) = conn
-            .query_row(
-                "SELECT status, last_error FROM notifications_outbox WHERE dedup_key = 'no-url:test'",
-                [],
-                |row| Ok((row.get(0)?, row.get(1)?)),
-            )
-            .unwrap();
-        assert_eq!(status, "dropped");
-        assert_eq!(error.as_deref(), Some("no_apprise_urls"));
-        let firing_count: i64 = conn
-            .query_row("SELECT COUNT(*) FROM notification_firings", [], |row| {
-                row.get(0)
-            })
-            .unwrap();
-        assert_eq!(firing_count, 0, "drops must not create firing history");
-    }
-}
+#[path = "dispatcher_tests.rs"]
+mod tests;
